Separate page and prompt contexts
