+++ frontend/src/components/ChatBot/__tests__/promptContext.test.ts
@@ -7,7 +7,6 @@ rstest.mock('config', () => ({
 }));
 
 import {
-  buildPageContext,
   buildPromptContext,
   buildPromptVariables,
   enrichPromptContext,
@@ -74,29 +73,6 @@ describe('formatHealthContext', () => {
 
   it('formats known health statuses', () => {
     expect(formatHealthContext('Degraded')).toBe(' with current health status Degraded');
-  });
-});
-
-describe('buildPageContext', () => {
-  it('does not append cluster in single-cluster environments', () => {
-    mockIsMultiCluster = false;
-    expect(buildPageContext('app', 'reviews', 'bookinfo', undefined, 'west')).toBe(
-      'User is seeing the information about application reviews of namespace bookinfo'
-    );
-  });
-
-  it('appends cluster in multi-cluster environments', () => {
-    mockIsMultiCluster = true;
-    expect(buildPageContext('app', 'reviews', 'bookinfo', undefined, 'west')).toBe(
-      'User is seeing the information about application reviews of namespace bookinfo in cluster west'
-    );
-    mockIsMultiCluster = false;
-  });
-  it('appends health status on detail views', () => {
-    mockIsMultiCluster = false;
-    expect(buildPageContext('service', 'details', 'bookinfo', undefined, undefined, 'Failure')).toBe(
-      'User is seeing the information about service details of namespace bookinfo with current health status Failure'
-    );
   });
 });
