+++ frontend/src/components/ChatBot/promptContext.ts
@@ -123,70 +123,6 @@ export const formatHealthContext = (status?: HealthStatusId): string => {
   return ` with current health status ${status}`;
 };
 
-export const buildPageContext = (
-  kind: string | undefined,
-  name: string | undefined,
-  namespace: string | undefined,
-  istio: string | undefined,
-  clusterName?: string | undefined,
-  healthStatus?: HealthStatusId
-): string | undefined => {
-  if (!kind) {
-    return undefined;
-  }
-
-  let context: string;
-
-  if (!name) {
-    switch (kind) {
-      case 'mesh':
-        context = 'User is seeing the mesh Graph';
-        break;
-      case 'graph':
-        context = 'User is seeing the Traffic Graph';
-        break;
-      case 'workloads':
-      case 'services':
-      case 'applications':
-      case 'namespaces':
-        context = `User is seeing the List of ${kind}`;
-        break;
-      case 'istio':
-        context = 'User is seeing the istio objects list';
-        break;
-      default:
-        context = `User is seeing the ${kind} page`;
-    }
-
-    if (namespace) {
-      context += ` for namespaces: ${namespace}`;
-    }
-  } else if (kind === 'istio') {
-    context = `User is seeing the istio object ${name} of namespace ${namespace} that is type ${istio || 'unknown'}`;
-  } else {
-    let kindStr = kind;
-    if (kind === 'app' || kind === 'applications') kindStr = 'application';
-    else if (kind === 'wk' || kind === 'workloads') kindStr = 'workload';
-    else if (kind === 'srv' || kind === 'services') kindStr = 'service';
-    else if (kind === 'namespace') kindStr = 'namespace';
-
-    context = `User is seeing the information about ${kindStr} ${name}`;
-    if (kind !== 'namespace' && namespace) {
-      context += ` of namespace ${namespace}`;
-    }
-  }
-
-  if (clusterName && isMultiCluster) {
-    context += ` in cluster ${clusterName}`;
-  }
-
-  if (name && healthStatus && healthStatus !== 'NA') {
-    context += formatHealthContext(healthStatus);
-  }
-
-  return context;
-};
-
 export const substitutePromptVariables = (text: string, variables: PromptVariables): string => {
   const merged = { ...PROMPT_VARIABLE_DEFAULTS, ...variables };
   return text.replace(/\{([a-z_]+)\}/g, (_match, key: string) => merged[key] ?? '');
